+++ CHANGES.rst
@@ -3,4 +3,4 @@ Change log
 
 v0.0.1
 ------
-* Beta of PyPAPER package implemented for rapid molecule similarity calculations.
+* Beta of roshambo package implemented for rapid molecule similarity calculations.